Add date and datetime scenarios to gigasecond
The gigasecond test suite is a bit odd in that there are two different
types for the same input.

Another oddity of the gigasecond test suite is that most of the tests don't actually
add any value. If you get one passing, you get them all passing.

This adds a date scenario to the inputs that have only a date, and datetime to the
ones that also include a time element.